+++ wagtailgeowidget/geocoders.py
@@ -1,2 +1,3 @@
 NOMINATIM = "nominatim"
 GOOGLE_MAPS = "google_maps"
+MAPBOX = "mapbox"