+++ wagtailgeowidget/static/wagtailgeowidget/js/geocoder-field-telepath.js
@@ -1,9 +1,10 @@
 // This file must follow ES5
 (function () {
-    function GeocoderFieldWrap(html, _id, geocoder, translations) {
+    function GeocoderFieldWrap(html, _id, geocoder, translations, params) {
         this.html = html;
         this.geocoder = geocoder;
         this.translations = translations;
+        this.params = params;
     }
 
     GeocoderFieldWrap.prototype.render = function (
@@ -18,11 +19,16 @@
         var fieldByGeocoder = {
             nominatim: NominatimGeocoderField,
             google_maps: GoogleMapsGeocoderField,
+            mapbox: MapboxGeocoderField,
         };
 
         var Field = fieldByGeocoder[this.geocoder];
 
-        var geocoderField = new Field({ id: id, translations: this.translations});
+        var geocoderField = new Field({
+            id: id,
+            translations: this.translations,
+            params: this.params,
+        });
         geocoderField.setState(initialState);
         return geocoderField;
     };